test: improve coverage for utility modules
Adds missing test coverage for critical utility functions across the
transactions module. Includes tests for:
- `utils.js` (currency conversion and CSV parsing)
- `dataLoader.js` (snapshot and realtime fetching fallbacks)
- `state.js` (edge cases in state management)
- `calculations.js` (invalid value parsing in CSV)
