--- fix_calcs.js
@@ -1,26 +1,0 @@
-const fs = require('fs');
-
-const path = 'tests/js/transactions/calculations.test.js';
-let content = fs.readFileSync(path, 'utf8');
-
-// I should see what parseCSV actually returns when they are empty strings. It passes the value directly:
-// const quantity = parseFloat(values[3]) || 0;
-// No, wait, if we look at parseCSV implementation in js/transactions/calculations.js:
-//             transactions.push({
-//                tradeDate: values[0],
-//                orderType: values[1],
-//                security: values[2],
-//                quantity: values[3],
-//                price: values[4],
-//                netAmount: Math.round(quantity * price).toString(),
-//                transactionId: transactionIdCounter,
-//            });
-// So quantity: values[3] is exactly what it is. If values[3] is "", then quantity is "". But netAmount is quantity * price rounded to string.
-// So if values[3] is "" and values[4] is "", quantity is "", price is "".
-// Let's modify the test to reflect what it ACTUALLY does.
-content = content.replace("expect(result[0].quantity).toBe('0');", "expect(result[0].quantity).toBe('');")
-                 .replace("expect(result[0].price).toBe('0');", "expect(result[0].price).toBe('');")
-                 .replace("expect(result[1].quantity).toBe('0');", "expect(result[1].quantity).toBe('invalid');")
-                 .replace("expect(result[1].price).toBe('0');", "expect(result[1].price).toBe('invalid');");
-
-fs.writeFileSync(path, content);
