--- test_utils.js
@@ -1,2 +1,0 @@
-import { convertBetweenCurrencies } from './js/transactions/utils.js';
-console.log(convertBetweenCurrencies(100, 'USD', '2024-01-01', 'EUR'));
