+++ tests/js/transactions/state.test.js
@@ -176,6 +176,12 @@ describe('state.js', () => {
     });
 
     test('setSelectedCurrency and getSelectedCurrency', () => {
+        setSelectedCurrency('');
+        expect(getSelectedCurrency()).toBe('USD');
+        setSelectedCurrency('');
+        expect(getSelectedCurrency()).toBe('USD');
+        setSelectedCurrency('');
+        expect(getSelectedCurrency()).toBe('USD');
         setSelectedCurrency(''); // Should ignore
         expect(getSelectedCurrency()).toBe('USD');
         setSelectedCurrency('CNY');
@@ -199,6 +205,15 @@ describe('state.js', () => {
         expect(getCompositionFilterTickers()).toEqual(['AAPL', 'MSFT']); // Deduplicates and cleans
 
         setCompositionFilterTickers(null);
+        expect(getCompositionFilterTickers()).toEqual([]);
+
+        setCompositionFilterTickers([]);
+        expect(getCompositionFilterTickers()).toEqual([]);
+
+        setCompositionFilterTickers([]);
+        expect(getCompositionFilterTickers()).toEqual([]);
+
+        setCompositionFilterTickers([]);
         expect(getCompositionFilterTickers()).toEqual([]);
 
         setCompositionFilterTickers([]);
